test(router): move the sync tpm regression into the existing tpm routing test file
tests/local_testing/test_tpm_rpm_routing_v2.py is the test file that already covers
LowestTPMLoggingHandler_v2, so the regression for the sync get_available_deployments
raise lives there instead of in a new file
